qa/requirements.txt
python-dotenv
langchain
openai
pypdf
chromadb
tiktoken
requests
beautifulsoup4
langchainhub
streamlit
semantic-router
langchain-community


#pysqlite3-binary
